pyproject.toml
[tool.poetry]
name = "iytdl"
version = "0.4.0"
description ="Aysnc Inline YouTube-DL for Pyrogram"
authors = ["Leorio Paradinight <62891774+code-rgb@users.noreply.github.com>"]
homepage = "https://github.com/iytdl/iytdl"
repository = "https://github.com/iytdl/iytdl"
keywords = ["iytdl", "yt-dlp", "youtube-dl", "pyrogram", "x"]
license = "GPLv3"
readme = "README.md"

classifiers = [
    "Development Status :: 4 - Beta",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
    "Topic :: Software Development",
]

[tool.poetry.dependencies]
python = ">=3.8,<4"
youtube-search-python = "1.4.6"
Pyrogram = "^1.2.9"
Pillow = "^8.3.1"
html-telegraph-poster = "^0.4.0"
hachoir = "^3.1.2"
mutagen = "^1.45.1"
aiosqlite = "^0.17.0"
aiohttp = "^3.7.4"
yt-dlp = "^2021.8.10"
TgCrypto = "^1.2.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
isort = "^5.7.0"
black = "^20.8b1"
autopep8 = "^1.5.7"
autoflake = "^1.4"

[tool.isort]
profile = "black"
include_trailing_comma = true
lines_after_imports = 2
lines_between_types = 1
use_parentheses = true
filter_files = true

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
